The Zenva turn based battler project
Some odd things happened as I followed the Enemy tutorial. The .pngs were not made into sprites when I dragged them into the assets folder - I had to manually set them as 2D sprites.
When I dragged the characters from the scene into the characters folder the icon on the character (in the folder) was skewed.